requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.0.1
django-address==0.2.8
django-authtools==1.7.0
django-cors-headers==3.3.0
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==9.2.0

python-decouple==3.6
pytz==2022.4
whitenoise==5.3.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
oauthlib==3.2.1
protobuf==4.21.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
psycopg2==2.8.6
psycopg2-binary==2.8.6
